feat: Step 1 구조 수정, Student 구조체 정의
main.swift에 있던 클래스와 구조체들을 별도의 팔일로 분리
Student 구조체 정의
